+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -6,13 +6,13 @@ import java.util.List;
 
 
 public interface ItemService {
-    public ItemDto createItem(ItemDto itemDto, int userId);
+    public ItemDto createItem(ItemDto itemDto, long userId);
 
-    public ItemDto updateItem(ItemDto itemDto, int userId);
+    public ItemDto updateItem(ItemDto itemDto, long userId);
 
-    public ItemDto getItemById(int itemId, int userId);
+    public ItemDto getItemById(long itemId, long userId);
 
-    public List<ItemDto> getItemsByUserId(int userId);
+    public List<ItemDto> getItemsByUserId(long userId);
 
     public List<ItemDto> getItemsByQuery(String query);
